Add hardware and software for reference
This is just some reference to the hardware setup that I'm using. It
also shows how I'm filling the BRAM up and then reading from it, which
is still just a proof of concept at this point.

I've also moved to the dedicated UART on the PI (ttyS0) and sped things
up to 230400 baud. All of this is still hard coded, for now.

Signed-off-by: Stephano Cetola <scetola@linuxfoundation.org>