add graphs and cleanup threading
The graphs are static for now, but at least they you have it.

The threading model was a little wonky. Basically we weren't using
gevent which gives us thread pools as a feature. Now we can get rid of
that pesky "background task" and just use threads like a normal human.

Signed-off-by: Stephano Cetola <scetola@linuxfoundation.org>